[1312] Provide the validation GraphQL datafetchers
Bug: https://github.com/eclipse-sirius/sirius-components/issues/1312
