--- packages/core/backend/sirius-components-graphql-utils/README.adoc
@@ -1,10 +1,0 @@
-= sirius-components-graphql-utils
-
-== Goal
-
-This project provides some utility services to help creating GraphQL schema and perform some queries.
-
-== Dependencies
-
-- GraphQL Java
-- Sirius Web Annotations
