--- packages/core/backend/sirius-components-graphql-utils/src/main/java/org/eclipse/sirius/components/graphql/utils/types/UploadScalarCoercing.java
@@ -1,42 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2019, 2020 Obeo.
- * This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v2.0
- * which accompanies this distribution, and is available at
- * https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *     Obeo - initial API and implementation
- *******************************************************************************/
-package org.eclipse.sirius.components.graphql.utils.types;
-
-import graphql.schema.Coercing;
-import graphql.schema.CoercingParseLiteralException;
-import graphql.schema.CoercingParseValueException;
-import graphql.schema.CoercingSerializeException;
-
-/**
- * The coercing policy for the upload scalar.
- *
- * @author sbegaudeau
- */
-public class UploadScalarCoercing implements Coercing<Object, Object> {
-
-    @Override
-    public Object serialize(Object dataFetcherResult) throws CoercingSerializeException {
-        return dataFetcherResult;
-    }
-
-    @Override
-    public Object parseValue(Object input) throws CoercingParseValueException {
-        return input;
-    }
-
-    @Override
-    public Object parseLiteral(Object input) throws CoercingParseLiteralException {
-        return input;
-    }
-
-}
