+++ packages/starters/backend/sirius-components-starter/pom.xml
@@ -117,6 +117,11 @@
 		</dependency>
 		<dependency>
 			<groupId>org.eclipse.sirius</groupId>
+			<artifactId>sirius-components-validation-graphql</artifactId>
+			<version>2022.7.1</version>
+		</dependency>
+		<dependency>
+			<groupId>org.eclipse.sirius</groupId>
 			<artifactId>sirius-components-emf</artifactId>
 			<version>2022.7.1</version>
 		</dependency>
