+++ packages/starters/backend/sirius-components-starter/src/main/java/org/eclipse/sirius/components/starter/SiriusWebStarterConfiguration.java
@@ -14,12 +14,20 @@ package org.eclipse.sirius.components.starter;
 
 import java.util.concurrent.Executors;
 
+import org.eclipse.sirius.components.collaborative.api.IEditingContextEventProcessorRegistry;
+import org.eclipse.sirius.components.collaborative.api.IRepresentationConfiguration;
+import org.eclipse.sirius.components.collaborative.api.IRepresentationEventProcessor;
 import org.eclipse.sirius.components.collaborative.api.ISubscriptionManagerFactory;
 import org.eclipse.sirius.components.collaborative.editingcontext.api.IEditingContextEventProcessorExecutorServiceProvider;
 import org.eclipse.sirius.components.collaborative.forms.WidgetSubscriptionManager;
 import org.eclipse.sirius.components.collaborative.forms.api.IWidgetSubscriptionManagerFactory;
 import org.eclipse.sirius.components.collaborative.representations.SubscriptionManager;
+import org.eclipse.sirius.components.core.api.IInput;
+import org.eclipse.sirius.components.core.api.IPayload;
+import org.eclipse.sirius.components.graphql.api.IEventProcessorSubscriptionProvider;
+import org.eclipse.sirius.components.graphql.api.IExceptionWrapper;
 import org.eclipse.sirius.components.graphql.ws.api.IGraphQLWebSocketHandlerListener;
+import org.eclipse.sirius.components.starter.services.ExceptionWrapper;
 import org.eclipse.sirius.components.web.concurrent.DelegatingRequestContextExecutorService;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
@@ -29,6 +37,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.socket.CloseStatus;
 import org.springframework.web.socket.TextMessage;
 import org.springframework.web.socket.WebSocketSession;
+
+import reactor.core.publisher.Flux;
 
 /**
  * Projects which depend on this starter project will automatically get all the components required to create a Sirius
@@ -97,5 +107,28 @@ public class SiriusWebStarterConfiguration {
                 // Do nothing
             }
         };
+    }
+
+    @Bean
+    @ConditionalOnMissingBean
+    public IEventProcessorSubscriptionProvider eventProcessorSubscriptionProvider(IEditingContextEventProcessorRegistry editingContextEventProcessorRegistry) {
+        return new IEventProcessorSubscriptionProvider() {
+            @Override
+            public <T extends IRepresentationEventProcessor> Flux<IPayload> getSubscription(String editingContextId, Class<T> representationEventProcessorClass,
+                    IRepresentationConfiguration representationConfiguration, IInput input) {
+                // @formatter:off
+                return editingContextEventProcessorRegistry.getOrCreateEditingContextEventProcessor(editingContextId)
+                        .flatMap(processor -> processor.acquireRepresentationEventProcessor(representationEventProcessorClass, representationConfiguration, input))
+                        .map(representationEventProcessor -> representationEventProcessor.getOutputEvents(input))
+                        .orElse(Flux.empty());
+                // @formatter:on
+            }
+        };
+    }
+
+    @Bean
+    @ConditionalOnMissingBean
+    public IExceptionWrapper exceptionWrapper() {
+        return new ExceptionWrapper();
     }
 }
