+++ packages/validation/backend/sirius-components-validation-graphql/.checkstyle
