+++ packages/validation/backend/sirius-components-validation-graphql/.settings/org.eclipse.core.resources.prefs
