+++ packages/validation/backend/sirius-components-validation-graphql/.settings/org.eclipse.jdt.core.prefs
