+++ packages/validation/backend/sirius-components-validation-graphql/.settings/org.eclipse.m2e.core.prefs
