+++ packages/validation/backend/sirius-components-validation-graphql/pom.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
- Copyright (c) 2019, 2022 Obeo.
+ Copyright (c) 2022 Obeo.
  This program and the accompanying materials
  are made available under the terms of the Eclipse Public License v2.0
  which accompanies this distribution, and is available at
@@ -13,7 +13,7 @@
 -->
 <project xmlns="http://maven.apache.org/POM/4.0.0"
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
 		<groupId>org.springframework.boot</groupId>
@@ -22,10 +22,10 @@
 		<relativePath /> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>org.eclipse.sirius</groupId>
-	<artifactId>sirius-components-graphql-api</artifactId>
+	<artifactId>sirius-components-validation-graphql</artifactId>
 	<version>2022.7.1</version>
-	<name>sirius-components-graphql-api</name>
-	<description>Sirius Components GraphQL API</description>
+	<name>sirius-components-validation-graphql</name>
+	<description>Sirius Components Validation GraphQL</description>
 
 	<properties>
 		<java.version>11</java.version>
@@ -47,8 +47,13 @@
 			<version>${graphql-java.version}</version>
 		</dependency>
 		<dependency>
-			<groupId>io.projectreactor</groupId>
-			<artifactId>reactor-core</artifactId>
+			<groupId>com.fasterxml.jackson.core</groupId>
+			<artifactId>jackson-databind</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.eclipse.sirius</groupId>
+			<artifactId>sirius-components-graphql-api</artifactId>
+			<version>2022.7.1</version>
 		</dependency>
 		<dependency>
 			<groupId>org.eclipse.sirius</groupId>
@@ -57,13 +62,25 @@
 		</dependency>
 		<dependency>
 			<groupId>org.eclipse.sirius</groupId>
-			<artifactId>sirius-components-core</artifactId>
+			<artifactId>sirius-components-collaborative-validation</artifactId>
 			<version>2022.7.1</version>
 		</dependency>
 		<dependency>
 			<groupId>org.eclipse.sirius</groupId>
-			<artifactId>sirius-components-collaborative</artifactId>
+			<artifactId>sirius-components-tests</artifactId>
 			<version>2022.7.1</version>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-test</artifactId>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.eclipse.sirius</groupId>
+			<artifactId>sirius-components-tests</artifactId>
+			<version>2022.7.1</version>
+			<scope>test</scope>
 		</dependency>
 	</dependencies>
