+++ packages/validation/backend/sirius-components-validation-graphql/src/test/resources/logback-test.xml
