+++ packages/web/backend/sirius-components-graphql/src/test/java/org/eclipse/sirius/components/graphql/controllers/GraphQLControllerTests.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019, 2020 Obeo.
+ * Copyright (c) 2019, 2022 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -18,7 +18,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.util.Map;
 
-import org.eclipse.sirius.components.graphql.utils.types.UploadScalarType;
+import org.eclipse.sirius.components.graphql.api.UploadScalarType;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
